binary greedy meshing

+++ VoxelEngine/Input.h
@@ -7,8 +7,8 @@ public:
 	static bool getKey(int key);
 	static bool getKeyDown(int key);
 	static bool getKeyUp(int key);
-	static double getMouseAxisX();
-	static double getMouseAxisY();
+	static float getMouseAxisX();
+	static float getMouseAxisY();
 private:
 	static void init();
 	static void update();
